feat(core,daemon): Phase 2a — FTS5 full-text search with IPC search method
Add upsert_fts/delete_fts/search_items to storage layer, wire IPC "search"
method, integrate FTS indexing into clipboard capture path, and fix isize
cast in clipboard.rs for Rust 1.89 compatibility.

+++ crates/copypaste-daemon/src/clipboard.rs
@@ -43,7 +43,7 @@ impl ClipboardMonitor {
 
             let (count, content) = unsafe {
                 let pb = NSPasteboard::generalPasteboard();
-                let count = pb.changeCount();
+                let count = pb.changeCount() as i64;
                 let s = pb.stringForType(NSPasteboardTypeString);
                 (count, s.map(|ns| ns.to_string()))
             };
